refactor: 🏗️ update project files structure
change:
 folder & files locations/layout.
