+++ src/Context/AppContext.js
@@ -1,0 +1,4 @@
+import { createContext } from 'react';
+
+const AppContext = createContext();
+export default AppContext;
