+++ src/Context/Provider.jsx
@@ -1,0 +1,23 @@
+import React, { useState, useMemo } from 'react';
+import PropTypes from 'prop-types';
+import AppContext from './AppContext';
+
+function AppProvider({ children }) {
+  const [collection, setCollection] = useState([]);
+
+  const context = useMemo(
+    () => ({
+      collection,
+      setCollection,
+    }),
+    [collection]
+  );
+
+  return <AppContext.Provider value={context}>{children}</AppContext.Provider>;
+}
+
+AppProvider.propTypes = {
+  children: PropTypes.node,
+}.isRequired;
+
+export default AppProvider;
