+++ src/Pages/App.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import logo from './logo.svg';
+import logo from '../Img/logo.svg';
 import './App.css';
 
 function App() {
